+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New release of the HyperDbg Debugger.
 - Export SDK API for starting a new process
 - Added the assembler commad 'a' for virtual memory ([link](https://docs.hyperdbg.org/commands/debugging-commands/a)) thanks to [@Abbas-MG](https://github.com/Abbas-MG)
 - Added the assembler commad '!a' for physical memory ([link](https://docs.hyperdbg.org/commands/extension-commands/a)) thanks to [@Abbas-MG](https://github.com/Abbas-MG)
+- Add and export SDK API for unsetting message callback
 
 ### Changed
 - Fix clearing '!monitor' hooks on a different process or if the process is closed (#409) ([link](https://github.com/HyperDbg/HyperDbg/issues/409))